Move message types to `message.ts`
