+++ src/webR/channel.ts
@@ -1,71 +1,14 @@
 import { AsyncQueue } from './queue';
-import { promiseHandles, newUUID, UUID, ResolveFn } from './utils';
+import { promiseHandles, ResolveFn } from './utils';
 import * as Synclink from 'synclink';
+import { Message,
+         newRequest,
+         Response,
+         SyncRequest } from './message';
 
 // FIXME: Why doesn't this work?
 // import { SynclinkTask } from 'synclink/task';
 import { SynclinkTask } from '../node_modules/synclink/dist/esm/task';
-
-
-export interface Message {
-  type: string;
-  data?: any;
-};
-
-export interface Request {
-  type: 'request';
-  data: {
-    uuid: UUID;
-    msg: Message;
-  }
-};
-
-export interface Response {
-  type: 'response';
-  data: {
-    uuid: UUID;
-    resp: any
-  }
-};
-
-export function newRequest(msg: Message,
-                           transferables?: [Transferable]): Request {
-  return newRequestResponseMessage(
-    {
-      type: 'request',
-      data: {
-        uuid: newUUID(),
-        msg: msg
-      }
-    },
-    transferables
-  )
-}
-
-export function newResponse(uuid: UUID,
-                            resp: any,
-                            transferables?: [Transferable]) : Response {
-  return newRequestResponseMessage(
-    {
-      type: 'response',
-      data: {
-        uuid,
-        resp
-      }
-    },
-    transferables
-  )
-}
-
-function newRequestResponseMessage<T>(msg: T,
-                                      transferables?: [Transferable]) : T {
-  // Signal to Synclink that the data contains objects we wish to
-  // transfer, as in `postMessage()`
-  if (transferables) {
-    Synclink.transfer(msg, transferables);
-  }
-  return msg;
-}
 
 
 // The channel structure is asymetric:
@@ -188,7 +131,6 @@ export function chanWorkerHandle(main: ChannelMain) {
 
 import { Endpoint } from './task-common';
 import { syncResponse } from './task-main';
-import { SyncRequest } from './message';
 
 const msgTypes = ['resolve', 'message', 'request', 'response', 'sync-request'];
